requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
paho-mqtt==1.6.1
scikit-learn==1.5.0
numpy==1.26.4
pytest==7.4.4
pytest-asyncio==0.23.7
httpx==0.27.0
